fix: 0 queries for some examples
The problem of NaNs in the loss was due to a 0-division. In particular, we divided summed scores by the number of query in the batch. The number of queries was zero for some samples in the dataset like the #4900222239. In this example we have the sentence "[/EN#216541/people Bicycler] performing [/EN#216542/other entertaining stunts] while [/EN#216552/other the sun] sets in [/EN#0/notvisual the distance] ." which contains a valid query, i.e. "Bicycler", but the word is out of vocab. Previous vocab system used the index 0 for out of vocab words, thus, the masking function counted that as padding. For this reason we introduced the <pad> token, among the <unk> token and set accordingly the word index.